restful web

+++ restTest/src/main/java/com/vpavlenko/restTest/messagedao/MessageDao.java
@@ -4,9 +4,12 @@ import com.vpavlenko.restTest.model.TableMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface MessageDao extends JpaRepository<TableMessage, Long> {
+public interface MessageDao extends JpaRepository<TableMessage, Integer> {
 
 
     TableMessage getByOrderNumber(int orderNumber);
+
 }
